% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_stat.R
\name{d_stat}
\alias{d_stat}
\title{calculate the energy statistics from balanced designed samples}
\usage{
d_stat(size, group, index, distance)
}
\arguments{
\item{size}{the sample size}

\item{group}{the number of the groups of the samples}

\item{index}{for permutation test, indicate how to divide the samples into groups}

\item{distance}{the distance matrix}
}
\value{
energy statistics used for multivariate test for equal distribution
}
\description{
This fucntion calculate the statistics to test multrivariate test for equal distributions based on energy method
}
\examples{
\dontrun{ 
pdim <- 10;delta <- 0.5;size = 30;group = 4
dat <- c(dat <- c(rt(n = size*pdim,df = 4,ncp = delta),
                  rt(n = size*pdim*(group - 1),df = 4))
dat <- matrix(dat,byrow = TRUE,nrow = total.size)
distance <- as.matrix(dist(dat,diag = TRUE,upper = TRUE,
                           method = "minkowski",p = 1))
index <- 1:size*group
d_stat(size,group,index,distance)
}
}
